refactor: change urls structure
