+++ backend/paul/urls_api.py
@@ -1,0 +1,6 @@
+from django.urls import include, path
+
+app_name = "api"
+urlpatterns = [
+    path("users/", include("users.urls_api")),
+]
