+++ backend/users/tests/test_api.py
@@ -13,7 +13,7 @@ class TestUserGetRolesApi(TestCase):
         self.client.login(username="testuser", password="testpass")
 
     def test_get_roles_success(self):
-        response = self.client.get(reverse("api:get-user-roles", kwargs={"user_id": self.user.pk}))
+        response = self.client.get(reverse("api:users:get-user-roles", kwargs={"user_id": self.user.pk}))
         self.assertEqual(response.status_code, 200)
         self.assertIn("roles", response.json())
         self.assertIsInstance(response.json()["roles"], list)
@@ -21,6 +21,6 @@ class TestUserGetRolesApi(TestCase):
     def test_get_roles_unauthenticated_redirect_to_login(self):
         self.client.logout()
 
-        api_url = reverse("api:get-user-roles", kwargs={"user_id": self.user.pk})
+        api_url = reverse("api:users:get-user-roles", kwargs={"user_id": self.user.pk})
         response = self.client.get(api_url)
         self.assertRedirects(response, f"{reverse('users:login')}?next={api_url}")
